Moved the payments utils test to unit tests
